requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
pywhatkit==5.4
gunicorn==21.2.0
